Refactor to db.clearDb
Refactor common logic from  user/clan models to database to be more expressive and maintainable.
